LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-invalidation-expected.txt

CONSOLE MESSAGE: Error: assert_not_equals: got disallowed value "784px"

FAIL CSS Values and Units Test: attr() invalidation Error: assert_not_equals: got disallowed value "784px"
